+++ OneBlink.SDK/models/FormsApp.cs
@@ -8,17 +8,17 @@ namespace OneBlink.SDK.Model
         public long id { get; set; }
         public string name { get; set; }
         public string slug { get; set; }
-        public string hostname { get; }
-        public DateTime createdAt { get; }
-        public DateTime updatedAt { get; }
+        public string hostname { get; set; }
+        public DateTime createdAt { get; set; }
+        public DateTime updatedAt { get; set; }
         public string organisationId { get; set; }
         public long formsAppEnvironmentId { get; set; }
-        public string oAuthClientId { get; }
+        public string oAuthClientId { get; set; }
         public List<long> formIds { get; set; }
-        public FormsAppStyles styles { get; }
+        public FormsAppStyles styles { get; set; }
         public FormsAppPWASettings pwaSettings { get; set; }
         public FormsAppWelcomeEmail welcomeEmail { get; set; }
-        public bool hasSamlIdentityProvider { get; }
+        public bool hasSamlIdentityProvider { get; set; }
         public List<string> notificationEmailAddresses { get; set; }
 
     }